+++ docs/CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         - Updated logic in `Confirm-XMLDataSet` and `Invoke-XMLDataCheck`
         - Removed private function `Get-XMLDataSet` for new private function `Get-RemoteFile`
     - Convert `1.5.0` or higher now required
+    - Added the `InsightView` parameter to both `Find-ModuleByCommand` and `Find-PSGModule`
+        - This provides a new results view that focuses on community insights
 - Build Updates
     - SSM Task now generates metadata file
     - InvokeBuild bumped from `5.10.2` to `5.10.3`